// src/music/artist/artistStats.dto.ts
export class ArtistStats{
    songNumber: number;
    totalPlays: number;
    avgPlaysPerDay: number;
    avgPlaysPerMonth: number;
// 
//        по периодам
// 

// по артисту в целом
    playsPerDay: any;
    playsPerMonth: any;

// по песням артиста
    songStats: any;
    // songPlays: any;
    songPlaysPerDay: any;
    // avgSongPlaysPerDay: any;
    songPlaysPerMonth: any;
    // avgSongPlaysPerMonth: any;

}